Fixed 3 warnings when compiling with g++.
git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@1577 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/scriptingclient.c
@@ -219,7 +219,7 @@ int main(int argc, char** argv)
                   else {
                      header = (const struct ScriptingCommonHeader*)buffer;
                      if( (rinfo.rinfo_ppid == ntohl(PPID_SP)) &&
-                         (received >= sizeof(struct ScriptingCommonHeader)) &&
+                         (received >= (ssize_t)sizeof(struct ScriptingCommonHeader)) &&
                          (ntohs(header->Length) == received) ) {
                         switch(header->Type) {
                            /* ====== Download message ==================== */
